\name{ssmn-package}
\alias{ssmn-package}
\docType{package}
\title{
Skew Scale Mixtures of Normal Distributions
}
\description{
It provides the density, distribution function, quantile function,
random number generator, likelihood function, direct and EM algorithm for Maximum Likelihood estimators for a given sample, all this for regression models using Skew Scale Mixtures of Normal Distributions.
}
\details{
\tabular{ll}{
Package: \tab ssmn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-08-08\cr
License: \tab GPL (>=2)\cr
}
}

\author{Clecio da Silva Ferreira \email{clecio.ferreira@ufjf.edu.br} and Luis Benites Sanchez <lbenitesanchez@gmail.com>
}

\references{
C.S. Ferreira; H. Bolfarine; V. H. Lachos (2011). Skew scale mixture of normal distributions: properties and estimation. Statistical Methodology, 8(2), 154-171.

C. S. Ferreira, V. H. Lachos, H. Bolfarine (2015). Inference and diagnostics in skew scale mixtures of normal regression models. Journal of Statistical Computation and Simulation, 85, 517-537.
}

\keyword{package}

\seealso{
\code{\link{ssmn}}
}

\examples{
#See examples linked above.


}
